use GasPriceMultiple

+++ eth/ethconfig/tx_pool.go
@@ -54,7 +54,7 @@ type DeprecatedTxPoolConfig struct {
 	// FreeClaimGasAddr is the address list for claim
 	FreeClaimGasAddr []string
 	// GasPriceMultiple is the factor claim tx gas price should mul
-	GasPriceMultiple float64
+	GasPriceMultiple uint64
 }
 
 // DeprecatedDefaultTxPoolConfig contains the default configurations for the transaction
